pyproject.toml
[project]
name = "projeto-mgi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncio>=3.4.3",
    "crawl4ai>=0.5.0.post4",
    "psycopg2-binary>=2.9.10",
    "ruff>=0.9.10",
]
